pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "uiao"
version = "0.3.0"
description = "UIAO — Unified Identity-Addressing-Overlay Architecture"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{ name = "WhalerMike" }]
dependencies = [
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "typer>=0.9",
  "rich>=13.0",
]

[project.scripts]
uiao = "uiao.cli:main"

# This pyproject covers only src/uiao/*. The existing impl/ subtree is
# installed separately via `pip install -e impl/` and contributes
# uiao.impl.* to the same PEP 420 `uiao` namespace at runtime.
[tool.setuptools.packages.find]
where = ["src"]
include = ["uiao*"]
namespaces = true

# Ship canon YAML / schemas / rules / KSI as package data so runtime
# code can access them via importlib.resources.
[tool.setuptools.package-data]
"uiao.canon" = ["**/*.yaml", "**/*.yml", "**/*.md", "**/*.json"]
"uiao.schemas" = ["**/*.json", "**/*.yaml", "**/*.yml"]
"uiao.rules" = ["**/*.yaml", "**/*.yml", "**/*.md"]
"uiao.ksi" = ["**/*.yaml", "**/*.yml", "**/*.json"]
